+++ CHANGELOG.md
@@ -84,6 +84,7 @@
 * nostr: add `CowTag` ([Yuki Kishimoto])
 * nostr: add `EventBorrow` ([Yuki Kishimoto])
 * nostr: add `HttpData::to_authorization` ([Yuki Kishimoto])
+* nostr: add `CoordinateBorrow` struct ([Yuki Kishimoto])
 * database: add `Events::first_owned` and `Events::last_owned` ([Yuki Kishimoto])
 * database: impl `FlatBufferDecodeBorrowed` for `EventBorrow` ([Yuki Kishimoto])
 * database: add `NostrDatabaseWipe` trait ([Yuki Kishimoto])